% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_survey_automation.R
\name{basic_survey_automation}
\alias{basic_survey_automation}
\title{Automating basic data}
\usage{
basic_survey_automation(survey_file, single_questions = NULL,
  multiple_questions = NULL, rank_questions = NULL)
}
\arguments{
\item{survey_file:}{Pollfish file to process}

\item{single_questions:}{Do you have a vector of single questions?}

\item{multiple_questions:}{Do you have a vector of multiple questions?}

\item{rank_questions:}{Do you have a vector of rank questions?}
}
\description{
Function that will run all crosstabs for all specified questions with basic demographic information as rows
}
\examples{
#Retrieve sample frequencies for rank choices
rank_choice <- c("Q7","Q8")
multiple_choice <- c("Q2", "Q9")
single_choice <-c("Q1","Q3","Q4","Q5","Q6", "Q11")
basic_survey_automation("Pollfish_Survey.xls",sinle_questions = single_choice,multiple_questions= multiple_choice,rank_questions = rank_choice)
}
\keyword{Pollfish,}
\keyword{basic}
\keyword{crosstab}
