% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_appendix_and_map.R
\name{create_appendix_and_map}
\alias{create_appendix_and_map}
\title{Creating an appendix and a map from a pollfish file}
\usage{
create_appendix_and_map(pollfish_file)
}
\arguments{
\item{pollfish_file:}{.xlsx file with the data}
}
\description{
Function that reads a Pollfish file and automatically creates frequency tables for the appendix and a map
}
\examples{
create_appendix_and_map("Pollfish_Survey.xls")
}
\keyword{Pollfish}
