% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map}
\alias{create_map}
\title{Create a map for respondent's state based on an appendix file}
\usage{
create_map(appendix_file)
}
\arguments{
\item{apendix_file:}{.xlsx file with appendix data}
}
\description{
Function that creates a map from an appendix file.
}
\examples{
create_map("appendix_file.xlsx")
}
\keyword{Pollfish}
