% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_crosstabs_by_variable.R
\name{multiple_crosstabs_by_variable}
\alias{multiple_crosstabs_by_variable}
\title{Creating multiple crosstabs from a pollfish file}
\usage{
multiple_crosstabs_by_variable(d_frame1, x_var1, multiple_choice_columns)
}
\arguments{
\item{d_frame1:}{Data frame created from a pollfish file}

\item{x_var1:}{Variable that will represent the rows}

\item{multiple_choice_columns:}{Vector with the code for all multiple choice variables}
}
\description{
Function that will run all multiple choice crosstabs for a single variable, save them in a list, and write them to file
}
\examples{
x <- read_pollfish_file("Pollfish_Survey.xls")
multiple_choice <- c("Q2", "Q9")
multiple_crosstabs_by_variable(x,"region", multiple_choice)
}
\keyword{Pollfish,}
\keyword{basic}
\keyword{crosstab}
