% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_question.R
\name{multiple_question}
\alias{multiple_question}
\title{Creating a basic crosstab for a multiple choice question from a pollfish file}
\usage{
multiple_question(d_frame, x_var, y_var)
}
\arguments{
\item{d_frame:}{Data frame created from a pollfish file}

\item{x_var:}{Variable that will represent the rows}

\item{y_var:}{Variable that will represent the columns}
}
\description{
Function that creates the crosstabulation between the x_var and the y_var for a single multiple choice question. Proportions displayed are row proportions and refer to % of total answers.
}
\examples{
x <- read_pollfish_file("Pollfish_Survey.xls")
multiple_question(x, "age", "Q2")
}
\keyword{Pollfish,}
\keyword{basic}
\keyword{crosstab}
