% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_whole_question.R
\name{rank_whole_question}
\alias{rank_whole_question}
\title{Creating rank crosstabs from a pollfish file}
\usage{
rank_whole_question(d_frame, x_var, rank_variables)
}
\arguments{
\item{d_frame1:}{Data frame created from a pollfish file}

\item{x_var:}{Variable that will represent the rows}

\item{rank_variables:}{Vector with the code for all rank questions}
}
\description{
Function that will run all rank questions crosstabs for a single variable, save them in a list, and write them to different files. There is a different file for each level of x_var
}
\examples{
x <- read_pollfish_file("Pollfish_Survey.xls")
rank_choice <- c("Q7","Q8")
rank_whole_question(x,"region", rank_choice)
}
\keyword{Pollfish,}
\keyword{basic}
\keyword{crosstab}
