% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pollfish_file.R
\name{read_pollfish_file}
\alias{read_pollfish_file}
\title{Reading a pollfish file}
\usage{
read_pollfish_file(enter_file)
}
\arguments{
\item{enter_file:}{.xlsx file with the data}
}
\description{
Function that reads a Pollfish file and automatically creates the region and income variables
}
\examples{
read_pollfish_file("Pollfish_Survey.xls")
}
\keyword{Pollfish}
