% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_sample_and_questions.R
\name{retrieve_sample_and_questions}
\alias{retrieve_sample_and_questions}
\title{Creating sample frequencies and question names from a pollfish file}
\usage{
retrieve_sample_and_questions(pollfish_file, column_names, rank = FALSE,
  prefix_for_file)
}
\arguments{
\item{pollfish_file:}{Pollfish file to process}

\item{column_names:}{Columns to pull information from. Do not mix question type}

\item{rank:}{Do you want to retrieve information for rank questions. Default is FALSE}

\item{prefix_for_file:}{Prefix for the .xlsx file created}
}
\description{
Function that will run all rank questions crosstabs for a single variable, save them in a list, and write them to different files. There is a different file for each level of x_var
}
\examples{
#Retrieve sample frequencies for rank choices
rank_choice <- c("Q7","Q8")
retrieve_sample_and_questions("Pollfish_Survey.xls",rank_choice,rank = TRUE,"rank")

#Retrieve sample frequencies for single choices
single_choice <-c("Q1","Q3","Q4","Q5","Q6", "Q11")
retrieve_sample_and_questions("Pollfish_Survey.xls",single_choice,rank = FALSE, "single")
}
\keyword{Pollfish,}
\keyword{basic}
\keyword{crosstab}
