% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_crosstab.R
\name{single_crosstab}
\alias{single_crosstab}
\title{Creating a single crosstab from a pollfish file}
\usage{
single_crosstab(d_frame, x_var, y_var)
}
\arguments{
\item{d_frame:}{Data frame created from a pollfish file}

\item{x_var:}{Variable that will represent the rows}

\item{y_var:}{Variable that will represent the columns}
}
\description{
Function that creates the crosstabulation between the x_var and the y_var. Proportions displayed are row proportions
}
\examples{
x <- read_pollfish_file("Pollfish_Survey.xls")
single_crosstab(x, "age", "Q4")
}
\keyword{Pollfish,}
\keyword{basic}
\keyword{crosstab}
