% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_crosstabs_by_variable.R
\name{single_crosstabs_by_variable}
\alias{single_crosstabs_by_variable}
\title{Creating multiple crosstabs from a pollfish file}
\usage{
single_crosstabs_by_variable(d_frame1, x_var1, single_choice_columns)
}
\arguments{
\item{d_frame1:}{Data frame created from a pollfish file}

\item{x_var1:}{Variable that will represent the rows}

\item{single_choice_columns:}{Vector with the code for all single choice variables}
}
\description{
Function that will run all single crosstabs for a single variable, save them in a list, and write them to file
}
\examples{
x <- read_pollfish_file("Pollfish_Survey.xls")
single_choice <-c("Q1","Q3","Q4","Q5","Q6", "Q11")
single_crosstabs_by_variable(x,"age",single_choice)
}
\keyword{Pollfish,}
\keyword{basic}
\keyword{crosstab}
